refactor: separate section interface from implementation in repo and server
